Fix linting error

+++ test/e2e/app/app.go
@@ -435,7 +435,7 @@ func (app *Application) PrepareProposal(
 		}
 		extCommitHex := hex.EncodeToString(extCommitBytes)
 		extTx := []byte(fmt.Sprintf("%s%d|%s", extTxPrefix, sum, extCommitHex))
-		extTxLen := int64(cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{extTx}))
+		extTxLen := cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{extTx})
 		app.logger.Info("preparing proposal with special transaction from vote extensions", "extTxLen", extTxLen)
 		if extTxLen > req.MaxTxBytes {
 			panic(fmt.Errorf("serious problem in the e2e app configuration; "+
@@ -457,10 +457,10 @@ func (app *Application) PrepareProposal(
 			app.logger.Error("detected tx that should not come from the mempool", "tx", tx)
 			continue
 		}
-		if totalBytes+int64(cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{tx})) > req.MaxTxBytes {
+		if totalBytes+cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{tx}) > req.MaxTxBytes {
 			break
 		}
-		totalBytes += int64(cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{tx}))
+		totalBytes += cmttypes.ComputeProtoSizeForTxs([]cmttypes.Tx{tx})
 		// Coherence: No need to call parseTx, as the check is stateless and has been performed by CheckTx
 		txs = append(txs, tx)
 	}
